Eliminate 'conversion may change sign of result' clang warnings in tests
Issue #558 (bdwgc).

* tests/disclaim_bench.c (main): Cast MS_TIME_DIFF() result to double
explicitly.
* tests/gctest.c [AT_END] (cons): Replace ~7 to ~7U.
* tests/leak.c (main): Change type of i local variable from int to
unsigned; remove cast of i to unsigned.
* tests/weakmap.c (weakmap_add, main): Cast weakobj_kind to int.

+++ tests/disclaim_bench.c
@@ -149,7 +149,7 @@ int main(int argc, char **argv)
         GC_gcollect();
 #       ifndef NO_CLOCK
             GET_TIME(tF);
-            t = MS_TIME_DIFF(tF, tI)*1e-3;
+            t = (double)MS_TIME_DIFF(tF, tI) * 1e-3;
 #       endif
 
 #       ifdef EMBOX
